+++ InvenShopfy.Core/Models/People/Supplier.cs
@@ -10,7 +10,7 @@ public class Supplier
     public string Name { get; set; } = String.Empty;
     public string PhoneNumber { get; set; } = String.Empty;
     public string Email { get; set; } = String.Empty;
-    public double SupplierCode { get; set; }
+    public long SupplierCode { get; set; }
     public string Country { get; set; } = String.Empty;
     public string City { get; set; } = String.Empty;
     public string Address { get; set; } = String.Empty;
